+++ OSXIcotools/common/strbuf.c
@@ -57,7 +57,7 @@ normalize_str_pos(const char *str, int32_t pos)
     if (str == NULL)
         return 0;
     if (pos >= 0)
-        return strnlen(str, pos);
+        return (int32_t)strnlen(str, pos);
     pos += 1 + strlen(str);
     if (pos >= 0)
         return pos;
